More rust alpha build fixes

+++ src/sync/sync.rs
@@ -138,7 +138,7 @@ mod platform {
     }
 }
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let program = &args[0];
 
     let options = [
